% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_objects.R
\name{Raw.keyvalues}
\alias{Raw.keyvalues}
\title{Raw.keyvalues Object}
\usage{
Raw.keyvalues()
}
\value{
Raw.keyvalues object
}
\description{
Raw.keyvalues Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Raw functions: \code{\link{Raw}}
}
