% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_objects.R
\name{Raw}
\alias{Raw}
\title{Raw Object}
\usage{
Raw(Raw.keyvalues = NULL, keyvalues = NULL)
}
\arguments{
\item{Raw.keyvalues}{The \link{Raw.keyvalues} object or list of objects}

\item{keyvalues}{No description}
}
\value{
Raw object
}
\description{
Raw Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Raw functions: \code{\link{Raw.keyvalues}}
}

