% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_objects.R
\name{Tokeninfo}
\alias{Tokeninfo}
\title{Tokeninfo Object}
\usage{
Tokeninfo(access_type = NULL, audience = NULL, email = NULL,
  email_verified = NULL, expires_in = NULL, issued_at = NULL,
  issued_to = NULL, issuer = NULL, nonce = NULL, scope = NULL,
  user_id = NULL, verified_email = NULL)
}
\arguments{
\item{access_type}{The access type granted with this token}

\item{audience}{Who is the intended audience for this token}

\item{email}{The email address of the user}

\item{email_verified}{Boolean flag which is true if the email address is verified}

\item{expires_in}{The expiry time of the token, as number of seconds left until expiry}

\item{issued_at}{The issue time of the token, as number of seconds}

\item{issued_to}{To whom was the token issued to}

\item{issuer}{Who issued the token}

\item{nonce}{Nonce of the id token}

\item{scope}{The space separated list of scopes granted to this token}

\item{user_id}{The obfuscated user id}

\item{verified_email}{Boolean flag which is true if the email address is verified}
}
\value{
Tokeninfo object
}
\description{
Tokeninfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
